EventType for FUB only
