+++ real_intent/deliver/utils.py
@@ -3,29 +3,8 @@ import requests
 import time
 import random
 from functools import wraps
-from enum import StrEnum
 
 from real_intent.internal_logging import log
-
-
-# ---- Models ----
-
-class EventType(StrEnum):
-    """Event types for adding a lead."""
-    REGISTRATION = "Registration"
-    INQUIRY = "Inquiry"
-    SELLER_INQUIRY = "Seller Inquiry"
-    PROPERTY_INQUIRY = "Property Inquiry"
-    GENERAL_INQUIRY = "General Inquiry"
-    VIEWED_PROPERTY = "Viewed Property"
-    SAVED_PROPERTY = "Saved Property"
-    VISITED_WEBSITE = "Visited Website"
-    INCOMING_CALL = "Incoming Call"
-    UNSUBSCRIBED = "Unsubscribed"
-    PROPERTY_SEARCH = "Property Search"
-    SAVED_PROPERTY_SEARCH = "Saved Property Search"
-    VISITED_OPEN_HOUSE = "Visited Open House"
-    VIEWED_PAGE = "Viewed Page"
 
 
 # ---- Parameterized Exceptions ----
